size_t

+++ config.h
@@ -13,11 +13,11 @@ typedef struct __attribute__((__packed__)) procinfo {
 	pid_t pid;
 	union {
 		char *cwd;
-		int cwd_rel;
+		size_t cwd_rel;
 	};
 	union {
 		char *command;
-		int command_rel;
+		size_t command_rel;
 	};
 } procinfo;
